% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{calculate_brier_score}
\alias{calculate_brier_score}
\title{Calculate Brier Score}
\usage{
calculate_brier_score(predicted_survival, time, status, time_point)
}
\arguments{
\item{predicted_survival}{Predicted survival probabilities at time_point (a vector)}

\item{time}{Survival times}

\item{status}{Event indicators}

\item{time_point}{Evaluation time point}
}
\value{
Brier Score value
}
\description{
Calculate Brier Score
}
