% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChekLR_BIC.R
\name{CheckLR_BIC}
\alias{CheckLR_BIC}
\title{Learning Rate Check and Selection based on BIC}
\usage{
CheckLR_BIC(
  primData,
  auxData,
  cov = c("X1", "X2"),
  statusvar = "status",
  lr_vec = seq(0.001, 0.004, by = 5e-04),
  l1 = 0.1,
  l2 = 0.1,
  nsteps = 200
)
}
\arguments{
\item{primData}{Target domain data.}

\item{auxData}{Source domain data.}

\item{cov}{Vector of covariate names.}

\item{statusvar}{Name of the status variable.}

\item{lr_vec}{Vector of candidate learning rate values.}

\item{l1}{lambda1 parameter (fixed).}

\item{l2}{lambda2 parameter (fixed).}

\item{nsteps}{Number of optimization steps.}
}
\value{
A list containing the optimal learning rate and BIC values.
}
\description{
Selects the optimal learning rate using the BIC criterion, while keeping other parameters fixed.
}
