% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaQ.R
\name{deltaQ}
\alias{deltaQ}
\title{Calculate Target Domain Cumulative Hazards}
\usage{
deltaQ(primData, q)
}
\arguments{
\item{primData}{Target domain data, must contain 'time' and 'status' columns.}

\item{q}{Hazard parameters estimated from source domain, containing 'breakPoints' and 'cumHazards'.}
}
\value{
A data frame containing dQ, cumQ, and t.
}
\description{
Calculates the cumulative hazard function for target domain data based on parameters estimated from the source domain.
}
