% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenSimData.R
\name{GenSimData}
\alias{GenSimData}
\title{Generate Simulated Survival Data}
\usage{
GenSimData(nprim = 200, naux = 500, setting = 1)
}
\arguments{
\item{nprim}{Number of samples in the target domain.}

\item{naux}{Number of samples in the source domain.}

\item{setting}{Simulation setting ID (1-4), controlling different parameter and distribution settings.}
}
\value{
A list containing target domain and source domain data.
}
\description{
Generates simulated survival data for TransCox model testing, including target and source domain data.
}
