% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBIC.R
\name{GetBIC}
\alias{GetBIC}
\title{Calculate Bayesian Information Criterion (BIC)}
\usage{
GetBIC(
  status,
  CovData,
  hazards,
  newBeta,
  newHaz,
  eta,
  xi,
  cutoff = 1e-05,
  lambda1 = NULL,
  lambda2 = NULL,
  lambda_beta = NULL
)
}
\arguments{
\item{status}{Vector of event status.}

\item{CovData}{Matrix of covariate data.}

\item{hazards}{Vector of hazard functions.}

\item{newBeta}{New regression coefficients.}

\item{newHaz}{New hazard functions.}

\item{eta}{Vector of eta parameters.}

\item{xi}{Vector of xi parameters.}

\item{cutoff}{Threshold for determining non-zero parameters.}

\item{lambda1}{L1 penalty parameter for eta (used for adaptive thresholding).}

\item{lambda2}{L1 penalty parameter for xi (used for adaptive thresholding).}

\item{lambda_beta}{L1 penalty parameter for beta_t (used for adaptive thresholding).}
}
\value{
BIC value.
}
\description{
Calculates the BIC value for the TransCox model, used for model selection and parameter tuning.
}
