% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCoverage.R
\name{GetCoverage}
\alias{GetCoverage}
\title{Calculate Confidence Interval Coverage}
\usage{
GetCoverage(estimate, trueBeta, SE)
}
\arguments{
\item{estimate}{Parameter estimate.}

\item{trueBeta}{True parameter value.}

\item{SE}{Standard error.}
}
\value{
0 or 1, indicating whether the true value is covered.
}
\description{
Checks if the 95\% confidence interval of the parameter estimate covers the true value.
}
