% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLogLike.R
\name{GetLogLike}
\alias{GetLogLike}
\title{Calculate Log-Likelihood Function}
\usage{
GetLogLike(status, CovData, hazards, newBeta, newHaz)
}
\arguments{
\item{status}{Vector of event status, supports 0/1 or 1/2 encoding.}

\item{CovData}{Matrix of covariate data.}

\item{hazards}{Vector of original hazard functions.}

\item{newBeta}{Vector of new regression coefficients.}

\item{newHaz}{Vector of new hazard functions.}
}
\value{
Log-likelihood value.
}
\description{
Calculates the log-likelihood value for the TransCox model, supporting different status encoding formats.
}
