% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPrimaryParam.R
\name{GetPrimaryParam}
\alias{GetPrimaryParam}
\title{Calculate Target Domain Specific Parameters}
\usage{
GetPrimaryParam(primData, q, estR)
}
\arguments{
\item{primData}{Target domain data, must contain 'time' and 'status' columns.}

\item{q}{Parameter vector estimated from source domain.}

\item{estR}{Regression coefficients estimated from source domain.}
}
\value{
A list containing target domain specific parameters.
}
\description{
Calculates target domain specific breakpoints and hazards for TransCox model parameter estimation.
}
