% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_lasso_model.R
\name{predict_cox_lasso}
\alias{predict_cox_lasso}
\title{Predict using Cox Lasso Model}
\usage{
predict_cox_lasso(model_fit, test_data, cov_names, use_1se = FALSE)
}
\arguments{
\item{model_fit}{Trained Cox Lasso model}

\item{test_data}{Test data}

\item{cov_names}{Covariate names}

\item{use_1se}{Whether to use lambda.1se instead of lambda.min}
}
\value{
Predicted risk scores
}
\description{
Predict using Cox Lasso Model
}
