% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTransCox_Sparse.R
\name{print.TransCox_Sparse}
\alias{print.TransCox_Sparse}
\title{Print TransCox_Sparse Results}
\usage{
\method{print}{TransCox_Sparse}(x, ...)
}
\arguments{
\item{x}{A TransCox_Sparse object.}

\item{...}{Additional arguments passed to print.}
}
\description{
Print method for TransCox_Sparse object.
}
