% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBtsp_transCox.R
\name{runBtsp_transCox}
\alias{runBtsp_transCox}
\title{Run Bootstrap TransCox Analysis}
\usage{
runBtsp_transCox(
  aData,
  pData,
  Tres,
  learning_rate = 0.002,
  nsteps = 200,
  BestLam1 = 0.1,
  BestLam2 = 0.1,
  nbootstrap = 100,
  trueBeta = c(-0.5, 0.5),
  Hrange = seq(0.15, 2, length.out = 50),
  cov = c("X1", "X2")
)
}
\arguments{
\item{aData}{Source domain data.}

\item{pData}{Target domain data.}

\item{Tres}{Time points.}

\item{learning_rate}{Learning rate.}

\item{nsteps}{Number of optimization steps.}

\item{BestLam1}{Optimal lambda1 parameter.}

\item{BestLam2}{Optimal lambda2 parameter.}

\item{nbootstrap}{Number of Bootstrap iterations.}

\item{trueBeta}{True parameter values (for evaluation).}

\item{Hrange}{Time range for cumulative hazard function evaluation.}

\item{cov}{Vector of covariate names.}
}
\value{
A list containing Bootstrap results.
}
\description{
Evaluates the stability and confidence intervals of the TransCox model using the Bootstrap method.
}
