% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTransCox_one.R
\name{runTransCox_one}
\alias{runTransCox_one}
\title{Run Single TransCox Model Training}
\usage{
runTransCox_one(
  Pout,
  l1 = 1,
  l2 = 1,
  learning_rate = 0.004,
  nsteps = 200,
  cov = c("X1", "X2")
)
}
\arguments{
\item{Pout}{Preprocessed result containing target domain data and source domain parameters.}

\item{l1}{L1 penalty parameter for eta.}

\item{l2}{L1 penalty parameter for xi.}

\item{learning_rate}{Learning rate.}

\item{nsteps}{Number of optimization steps.}

\item{cov}{Vector of covariate names.}
}
\value{
A list containing eta and xi.
}
\description{
Runs a single iteration of TransCox model training using the Python backend to estimate eta and xi parameters.
}
