% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelLR_By_BIC.R
\name{SelLR_By_BIC}
\alias{SelLR_By_BIC}
\title{Joint Selection of Learning Rate and Steps Based on BIC}
\usage{
SelLR_By_BIC(
  primData,
  auxData,
  cov = c("X1", "X2"),
  statusvar = "status",
  lambda1 = 0.1,
  lambda2 = 0.1,
  learning_rate_vec = c(1e-04, 5e-04, 0.001, 0.002, 0.003, 0.004, 0.005),
  nsteps_vec = c(100, 200, 300, 400)
)
}
\arguments{
\item{primData}{Target domain data.}

\item{auxData}{Source domain data.}

\item{cov}{Vector of covariate names.}

\item{statusvar}{Name of the status variable.}

\item{lambda1}{Fixed lambda1 parameter.}

\item{lambda2}{Fixed lambda2 parameter.}

\item{learning_rate_vec}{Vector of candidate learning rates.}

\item{nsteps_vec}{Vector of candidate optimization steps.}
}
\value{
A list containing the optimal learning rate, steps, and BIC matrix.
}
\description{
Selects the optimal learning rate and number of optimization steps simultaneously using the BIC criterion.
}
